requirements.txt
google-genai>=1.0.0
gradio>=5.0.0
pydantic>=2.0.0
numpy>=1.26.0
scipy>=1.13.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
